separate BLList and BLDict into colls.py
